+++ nodes/utils/hueUtils.js
@@ -3,13 +3,7 @@
 // const hueApi = require('node-hue-api')
 const hueApiV2 = require('node-hue')
 const { EventEmitter } = require('events')
-
-
 const https = require('https');
-
-
-
-
 
 class classHUE extends EventEmitter {
   constructor(_hueBridgeIP, _username, _clientkey, _bridgeid) {
@@ -25,70 +19,96 @@ class classHUE extends EventEmitter {
     this.commandQueue = []
     this.closePushEventStream = false
     this.timerwriteQueueAdd = setTimeout(this.handleQueue, 3000) // First start
-
+    //this.run()
     // start the SSE Stream Receiver
     // #############################################
-    const options = {
-      host: _hueBridgeIP, // Indirizzo IP del tuo bridge Philips Hue
-      path: '/eventstream/clip/v2', // Il percorso dell'API per gli eventi
-      method: 'GET',
-      headers: {
-        'Connection': 'keep-alive',
-        'hue-application-key': _username
-        //'Accept': 'text/event-stream'
-      },
-      rejectUnauthorized: false
-    }
+    // const options = {
+    //   host: _hueBridgeIP, // Indirizzo IP del tuo bridge Philips Hue
+    //   path: '/eventstream/clip/v2', // Il percorso dell'API per gli eventi
+    //   method: 'GET',
+    //   headers: {
+    //     'Connection': 'keep-alive',
+    //     'hue-application-key': _username
+    //     //'Accept': 'text/event-stream'
+    //   },
+    //   rejectUnauthorized: false
+    // }
 
-    // Funzione per la gestione della risposta
-    const handleResponse = (response) => {
-      let data = '';
+    // // Funzione per la gestione della risposta
+    // const handleResponse = (response) => {
+    //   let data = '';
 
-      response.on('data', (chunk) => {
-        data += chunk;
-      });
+    //   response.on('data', (chunk) => {
+    //     data += chunk;
+    //   });
 
-      response.on('end', () => {
-        try {
-          const events = JSON.parse(data)
-          // An array event "Container", can have multiple events.
-          // for..loop is more efficent. We need speed.
-          for (let index = 0; index < events.length; index++) {
-            const oEvento = events[index]
-            if (oEvento.type === 'update') {
-              for (let i = 0; i < oEvento.data.length; i++) {
-                const element = oEvento.data[i]
-                this.emit('event', element)
-              }
-            }
-          }
-        } catch (error) {
-          console.log('KNXUltimateHUEConfig: classHUE: response.on(end): ' + error.message)
+    //   response.on('end', () => {
+    //     try {
+    //       const events = JSON.parse(data)
+    //       // An array event "Container", can have multiple events.
+    //       // for..loop is more efficent. We need speed.
+    //       for (let index = 0; index < events.length; index++) {
+    //         const oEvento = events[index]
+    //         if (oEvento.type === 'update') {
+    //           for (let i = 0; i < oEvento.data.length; i++) {
+    //             const element = oEvento.data[i]
+    //             this.emit('event', element)
+    //           }
+    //         }
+    //       }
+    //     } catch (error) {
+    //       console.log('KNXUltimateHUEConfig: classHUE: response.on(end): ' + error.message)
+    //     }
+    //     req();
+    //   });
+    // };
+    // // Funzione per richiedere gli eventi
+    // const req = () => {
+    //   if (this.closePushEventStream) return // I'm destroying the class
+    //   const request = https.request(options, handleResponse);
+    //   request.on('error', (error) => {
+    //     console.log('KNXUltimateHUEConfig: classHUE: request.on(error): ' + error.message)
+    //     // Restart the connection
+    //     setTimeout(() => {
+    //       this.commandQueue = []
+    //       req();
+    //     }, 2000);
+    //   });
+    //   request.end();
+    // };
+
+    // // Starts the connection for the first time
+    // req();
+
+    // Eventstream Reader using hueApiV2
+    const runStreamReader = async () => {
+      try {
+        const listener = (event) => {
+          //console.log(event)
+          event.data.forEach(element => {
+            if (event.type === 'update') this.emit('event', element)
+          });        
         }
-        req();
-      });
-    };
-    // Funzione per richiedere gli eventi
-    const req = () => {
-      if (this.closePushEventStream) return // I'm destroying the class
-      const request = https.request(options, handleResponse);
-      request.on('error', (error) => {
-        console.log('KNXUltimateHUEConfig: classHUE: request.on(error): ' + error.message)
-        // Restart the connection
-        setTimeout(() => {
-          this.commandQueue = []
-          req();
-        }, 2000);
-      });
-      request.end();
-    };
+        const hueEventStream = hueApiV2.connect({
+          host: this.hueBridgeIP,
+          key: this.username,
+          eventListener: listener
+        })
+      } catch (error) {
+        console.log('KNXUltimateHUEConfig: classHUE: const run = async: ' + error.message)
+      }
+    }
+    runStreamReader()
 
-    // Starts the connection for the first time
-    req();
+
+
+    // #############################################
   }
-// #############################################
-  
-  
+
+
+
+
+
   // Handle the send queue
   // ######################################
   handleQueue = async () => {
@@ -170,8 +190,6 @@ class classHUE extends EventEmitter {
       return ({ devices: error.message })
     }
   }
-
-
 
 
   close = async () => {